+++ src/vmm/src/utilities/test_utils/mod.rs
@@ -6,7 +6,6 @@ use std::sync::{Arc, Mutex};
 
 use crate::builder::build_microvm_for_boot;
 use crate::resources::VmResources;
-use crate::seccomp_filters::{get_filters, SeccompConfig};
 use crate::utilities::mock_resources::{MockBootSourceConfig, MockVmConfig, MockVmResources};
 use crate::vmm_config::boot_source::BootSourceConfig;
 use crate::vmm_config::instance_info::InstanceInfo;
@@ -14,7 +13,6 @@ use crate::{EventManager, Vmm};
 
 pub fn create_vmm(_kernel_image: Option<&str>, is_diff: bool) -> (Arc<Mutex<Vmm>>, EventManager) {
     let mut event_manager = EventManager::new().unwrap();
-    let empty_seccomp_filters = get_filters(SeccompConfig::None).unwrap();
 
     let boot_source_cfg = MockBootSourceConfig::new().with_default_boot_args();
     #[cfg(target_arch = "aarch64")]
@@ -38,7 +36,7 @@ pub fn create_vmm(_kernel_image: Option<&str>, is_diff: bool) -> (Arc<Mutex<Vmm>
             &InstanceInfo::default(),
             &resources,
             &mut event_manager,
-            &empty_seccomp_filters,
+            None,
         )
         .unwrap(),
         event_manager,